+++ projects/aca-about/src/test.ts
@@ -43,7 +43,7 @@ declare const require: {
 
 // First, initialize the Angular testing environment.
 getTestBed().initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting(), {
-    teardown: { destroyAfterEach: false }
+  teardown: { destroyAfterEach: false }
 });
 // Then we find all the tests.
 const context = require.context('./', true, /\.spec\.ts$/);
