+++ projects/aca-shared/store/src/actions/app-action-types.ts
@@ -23,33 +23,24 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from './actions/app-action-types';
-export * from './actions/context-menu-action-types';
-export * from './actions/router-action-types';
-export * from './actions/template-action-types';
-export * from './actions/app.actions';
-export * from './actions/library.actions';
-export * from './actions/node.actions';
-export * from './actions/router.actions';
-export * from './actions/search.actions';
-export * from './actions/snackbar.actions';
-export * from './actions/upload.actions';
-export * from './actions/viewer.actions';
-export * from './actions/metadata-aspect.actions';
-export * from './actions/template.actions';
-export * from './actions/contextmenu.actions';
-
-export * from './effects/dialog.effects';
-export * from './effects/router.effects';
-export * from './effects/snackbar.effects';
-
-export * from './models/delete-status.model';
-export * from './models/deleted-node-info.model';
-export * from './models/node-info.model';
-export * from './models/search-option.model';
-
-export * from './selectors/app.selectors';
-
-export * from './states/app.state';
-
-export * from './store.module';
+export enum AppActionTypes {
+  SetSettingsParameter = 'SET_SETTINGS_PARAMETER',
+  SetInitialState = 'SET_INITIAL_STATE',
+  SetHeaderColor = 'SET_HEADER_COLOR',
+  SetCurrentFolder = 'SET_CURRENT_FOLDER',
+  SetCurrentVersion = 'SET_CURRENT_VERSION',
+  SetCurrentUrl = 'SET_CURRENT_URL',
+  SetUserProfile = 'SET_USER_PROFILE',
+  SetRepositoryInfo = 'SET_REPOSITORY_INFO',
+  ToggleInfoDrawer = 'TOGGLE_INFO_DRAWER',
+  ToggleDocumentDisplayMode = 'TOGGLE_DOCUMENT_DISPLAY_MODE',
+  Logout = 'LOGOUT',
+  ReloadDocumentList = 'RELOAD_DOCUMENT_LIST',
+  ResetSelection = 'RESET_SELECTION',
+  SetInfoDrawerState = 'SET_INFO_DRAWER_STATE',
+  SetInfoDrawerMetadataAspect = 'SET_INFO_DRAWER_METADATA_ASPECT',
+  CloseModalDialogs = 'CLOSE_MODAL_DIALOGS',
+  SetFileUploadingDialog = 'SET_FILE_UPLOADING_DIALOG',
+  ShowInfoDrawerPreview = 'SHOW_INFO_DRAWER_PREVIEW',
+  SetInfoDrawerPreviewState = 'SET_INFO_DRAWER_PREVIEW_STATE'
+}
