+++ projects/adf-office-services-ext/src/test.ts
@@ -34,7 +34,7 @@ declare const require: any;
 
 // First, initialize the Angular testing environment.
 getTestBed().initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting(), {
-    teardown: { destroyAfterEach: false }
+  teardown: { destroyAfterEach: false }
 });
 // Then we find all the tests.
 const context = require.context('./', true, /\.spec\.ts$/);
